feat(dir): moved model class to appropriate directories
