+++ app/src/main/java/com/example/covivid/Utils/Common.java
@@ -5,7 +5,7 @@ import android.view.WindowManager;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.covivid.Model.BaseCovidReport;
+import com.example.covivid.Model.CovidReport.BaseCovidReport;
 import com.example.covivid.Retrofit.ICovidAPI;
 import com.example.covivid.Retrofit.RetrofitClient;
